# server/swagger.yaml
openapi: 3.0.1
info:
  title: Ecommerce API
  version: 1.0.0
  description: 'Basic API targetting commerce sector '
paths:
...
  /users/{userid}:
    summary: 'Related with username and email '
    description: GET and PUT operations
    parameters:
      - in: path
        name: userid
        schema:
          type: integer
        required: true
    get:
      summary: get user
      description: Retrieves username and email from authenticated user
      operationId: get-user
      responses:
        '200':
          description: Success
          content:
            application/json:
              examples:
                user retrieval:
                  value: |-
                    {
                        "id": 25,
                        "username": "Jdoe1989",
                        "email": "jdoe@gmail.com"
                    }
        '401':
          description: Unauthorized Access
          content:
            application/json:
              examples:
                Unauthenticated:
                  value: |-
                    {
                        "message": "User Not Logged In"
                    }
        '403':
          description: Forbidden Access
          content:
            application/json:
              examples:
                Forbidden:
                  value: |-
                    {
                        "message": "Unauthorized Access"
                    }
        '500':
          description: Server Error
          content:
            application/json:
              examples:
                Server Error:
                  value: |-
                    {
                      "Message": "Error Message"
                    }
      tags:
        - Users
    put:
      summary: Update username and/or email
      description: PUT operation on username and/or email
      operationId: update-user
      requestBody:
        required: true
        content:
          aplication/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  minLength: 4
                  maxLength: 40
                  description: Must contain alphanumeric values
                  example: Jdoe1989
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Success
          content:
            application/json:
              examples:
                updated username and/or password:
                  value: |-
                    {
                        "username": "updatedusername",
                        "email": "updatedpassword"
                    }
        '401':
          description: Unauthorized Access
          content:
            application/json:
              examples:
                Error Response:
                  value: |-
                    {
                        "message": "User Not Logged In"
                    }
        '403':
          description: Forbidden Access
          content:
            application/json:
              examples:
                Error response:
                  value: |-
                    {
                        "message": "Unauthorized Access"
                    }
        '409':
          description: Conflict, can also happend if user is already in  Use
          content:
            application/json:
              examples:
                Error Response:
                  value: |-
                    {
                        "message": "Email already in Use"
                    }
        '422':
          description: >-
            Unprocessable Entity, meaning one or more field failed  validation
            process.
          content:
            application/json:
              examples:
                Error Response:
                  value: |-
                    {
                        "Errors": {
                             "email": {
                                "value": "jdoe@gmail",
                                "msg": "Invalid value",
                                "param": "email",
                                "location": "body"
                            }
                        }
                    }
        '500':
          description: Server Error
          content:
            application/json:
              examples:
                Server Error:
                  value: |-
                    {
                      "Message": "Error Message"
                    }
      tags:
        - Users
...